+++ changelogs/mediation/digitalturbine/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to Digital Turbine Adapter for Yandex Mobile Ads Mediation will be documented in this file.
 
+## Version 8.3.7.2
+
+#### Updated
+* Yandex Mobile Ads SDK version 7.16.0
+
 ## Version 8.3.7.1
 
 #### Updated