+++ changelogs/mediation/startapp/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to StartApp Adapter for Yandex Mobile Ads Mediation will be documented in this file.
 
+## Version 5.0.2.16
+
+#### Updated
+* Yandex Mobile Ads SDK version 7.16.0
+
 ## Version 5.0.2.15
 
 #### Updated