+++ changelogs/mediation/unityads/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to UnityAds Adapter for Yandex Mobile Ads Mediation will be documented in this file.
 
+## Version 4.15.1.2
+
+#### Updated
+* Yandex Mobile Ads SDK version 7.16.0
+
 ## Version 4.15.1.1
 
 #### Updated