+++ changelogs/mobileads/CHANGELOG.md
@@ -2,9 +2,19 @@
 
 All notable changes to Yandex Mobile Ads SDK will be documented in this file.
 
+## Version 7.16.0
+
+### Added
+* Improvements and optimizations
+* Update AppMetrica SDK 7.12
+
 ## Version 7.15.2
 
 ### Added
+* Added new InStream API (experimental)
+* Improved ad validation
+
+### Fixed
 * Improvements and optimizations
 
 ## Version 7.15.1